+++ crates/plugin-zustand/src/lib.rs
@@ -2,6 +2,10 @@
 //
 // Check the `codegen` command in the `tauri-store-cli` crate.
 // https://github.com/ferreira-tb/tauri-store/tree/main/crates/tauri-store-cli
+//
+// To modify the behavior of the plugin, you must either change the
+// upstream `tauri-store` crate or update the code generation itself.
+// This ensures that all plugins maintain consistent behavior.
 
 #![forbid(unsafe_code)]
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
@@ -12,12 +16,12 @@ mod command;
 mod manager;
 mod zustand;
 
+use serde::de::DeserializeOwned;
 use std::collections::HashSet;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 use std::time::Duration;
-use tauri::plugin::TauriPlugin;
+use tauri::plugin::{PluginApi, TauriPlugin};
 use tauri::{AppHandle, RunEvent, Runtime};
-use tauri_store::CollectionBuilder;
 
 #[cfg(feature = "unstable-migration")]
 use tauri_store::Migrator;
@@ -29,16 +33,15 @@ pub use zustand::{Zustand, ZustandMarker};
 #[cfg(feature = "unstable-migration")]
 pub use tauri_store::{Migration, MigrationContext};
 
-// The `CollectionBuilder` macro depends on this.
-type Marker = ZustandMarker;
+#[cfg(target_os = "ios")]
+tauri::ios_plugin_binding!(init_plugin_zustand_plugin);
 
 /// Builder for the Zustand plugin.
-#[derive(CollectionBuilder)]
 pub struct Builder<R: Runtime> {
   path: Option<PathBuf>,
   default_save_strategy: SaveStrategy,
   autosave: Option<Duration>,
-  on_load: Option<Box<OnLoadFn<R, Marker>>>,
+  on_load: Option<Box<OnLoadFn<R, ZustandMarker>>>,
   pretty: bool,
   save_denylist: HashSet<StoreId>,
   sync_denylist: HashSet<StoreId>,
@@ -48,10 +51,146 @@ pub struct Builder<R: Runtime> {
 }
 
 impl<R: Runtime> Builder<R> {
+  /// Creates a new builder instance with default values.
+  pub fn new() -> Self {
+    Self::default()
+  }
+
+  /// Sets the autosave interval for all stores.
+  #[must_use]
+  pub fn autosave(mut self, interval: Duration) -> Self {
+    self.autosave = Some(interval);
+    self
+  }
+
+  /// Sets the default save strategy to be used by the stores.
+  #[must_use]
+  pub fn default_save_strategy(mut self, strategy: SaveStrategy) -> Self {
+    self.default_save_strategy = strategy;
+    self
+  }
+
+  /// Registers a closure to be called when a store is loaded.
+  #[must_use]
+  pub fn on_load<F>(mut self, f: F) -> Self
+  where
+    F: Fn(&Store<R, ZustandMarker>) -> Result<()> + Send + Sync + 'static,
+  {
+    self.on_load = Some(Box::new(f));
+    self
+  }
+
+  /// Directory where the stores will be saved.
+  #[must_use]
+  pub fn path(mut self, path: impl AsRef<Path>) -> Self {
+    let path = path.as_ref().to_path_buf();
+    self.path = Some(path);
+    self
+  }
+
+  /// Sets whether the store files should be pretty printed.
+  #[must_use]
+  pub fn pretty(mut self, yes: bool) -> Self {
+    self.pretty = yes;
+    self
+  }
+
+  /// Sets a list of stores that should not be saved to disk.
+  #[must_use]
+  pub fn save_denylist<I, T>(mut self, denylist: I) -> Self
+  where
+    I: IntoIterator<Item = T>,
+    T: AsRef<str>,
+  {
+    self.save_denylist.extend(
+      denylist
+        .into_iter()
+        .map(|it| StoreId::from(it.as_ref())),
+    );
+
+    self
+  }
+
+  /// Sets a list of stores that should not be synchronized across windows.
+  #[must_use]
+  pub fn sync_denylist<I, T>(mut self, denylist: I) -> Self
+  where
+    I: IntoIterator<Item = T>,
+    T: AsRef<str>,
+  {
+    self.sync_denylist.extend(
+      denylist
+        .into_iter()
+        .map(|it| StoreId::from(it.as_ref())),
+    );
+
+    self
+  }
+
+  /// Defines a migration for a store.
+  #[must_use]
+  #[cfg(feature = "unstable-migration")]
+  pub fn migration(mut self, id: impl Into<StoreId>, migration: Migration) -> Self {
+    self.migrator.add_migration(id.into(), migration);
+    self
+  }
+
+  /// Defines multiple migrations for a store.
+  #[must_use]
+  #[cfg(feature = "unstable-migration")]
+  pub fn migrations<I>(mut self, id: impl Into<StoreId>, migrations: I) -> Self
+  where
+    I: IntoIterator<Item = Migration>,
+  {
+    self
+      .migrator
+      .add_migrations(id.into(), migrations);
+
+    self
+  }
+
+  /// Sets a closure to be called before each migration step.
+  #[must_use]
+  #[cfg(feature = "unstable-migration")]
+  pub fn on_before_each_migration<F>(mut self, f: F) -> Self
+  where
+    F: Fn(MigrationContext) + Send + Sync + 'static,
+  {
+    self.migrator.on_before_each(f);
+    self
+  }
+
+  fn build_collection(self, handle: Handle<R>) -> Result<()> {
+    let mut builder = StoreCollection::<R, ZustandMarker>::builder()
+      .default_save_strategy(self.default_save_strategy)
+      .pretty(self.pretty)
+      .save_denylist(&self.save_denylist)
+      .sync_denylist(&self.sync_denylist);
+
+    if let Some(path) = self.path {
+      builder = builder.path(path);
+    }
+
+    if let Some(on_load) = self.on_load {
+      builder = builder.on_load(on_load);
+    }
+
+    if let Some(duration) = self.autosave {
+      builder = builder.autosave(duration);
+    }
+
+    #[cfg(feature = "unstable-migration")]
+    {
+      builder = builder.migrator(self.migrator);
+    }
+
+    builder.build(handle, env!("CARGO_PKG_NAME"))
+  }
+
   /// Builds the Zustand plugin.
   pub fn build(self) -> TauriPlugin<R> {
     tauri::plugin::Builder::new("zustand")
-      .setup(|app, _| setup(app, self))
+      .setup(|app, api| setup(app, api, self))
       .on_event(on_event)
       .invoke_handler(tauri::generate_handler![
         command::allow_save,
@@ -83,11 +222,46 @@ impl<R: Runtime> Builder<R> {
   }
 }
 
-fn setup<R>(app: &AppHandle<R>, builder: Builder<R>) -> BoxResult<()>
+impl<R: Runtime> Default for Builder<R> {
+  fn default() -> Self {
+    Self {
+      path: None,
+      default_save_strategy: SaveStrategy::default(),
+      autosave: None,
+      on_load: None,
+      pretty: false,
+      save_denylist: HashSet::default(),
+      sync_denylist: HashSet::default(),
+
+      #[cfg(feature = "unstable-migration")]
+      migrator: Migrator::default(),
+    }
+  }
+}
+
+#[cfg(any(target_os = "windows", target_os = "linux", target_os = "macos"))]
+fn setup<R, D>(app: &AppHandle<R>, _api: PluginApi<R, D>, builder: Builder<R>) -> BoxResult<()>
 where
   R: Runtime,
+  D: DeserializeOwned,
 {
-  builder.build_collection(app)?;
+  let handle = Handle::new(app.clone());
+  builder.build_collection(handle)?;
+  Ok(())
+}
+
+#[cfg(any(target_os = "android", target_os = "ios"))]
+fn setup<R, D>(app: &AppHandle<R>, api: PluginApi<R, D>, builder: Builder<R>) -> BoxResult<()>
+where
+  R: Runtime,
+  D: DeserializeOwned,
+{
+  #[cfg(target_os = "android")]
+  let handle = api.register_android_plugin("", "ZustandPlugin")?;
+  #[cfg(target_os = "ios")]
+  let handle = api.register_ios_plugin(init_plugin_zustand_plugin)?;
+
+  builder.build_collection(Handle::new(handle))?;
   Ok(())
 }